successfully sign the claim
