+++ lib/js_service_encointer/src/service/account.js
@@ -7,6 +7,7 @@ import { hexToU8a, u8aToHex, hexToString } from '@polkadot/util';
 import generateIcon from '@polkadot/ui-shared/icons';
 
 import { Keyring } from '@polkadot/keyring';
+import { createType } from '@polkadot/types';
 const keyring = new Keyring({ ss58Format: 0, type: 'sr25519' });
 
 function gen () {
@@ -257,6 +258,40 @@ function sendTx (txInfo, paramList) {
   });
 }
 
+/**
+ * Signs the claimHex with the provided key.
+ *
+ * Note: Even though this is an Encointer function, I put it here as the keys need to be initialized via the initKey()
+ * method, whose scope is apparently restricted to this module.
+ * @param claimHex
+ * @param pubKey
+ * @param password
+ * @returns {Promise<{Attestation, String}>} Runtime attestation struct and its hex Value
+ */
+export async function attestClaimOfAttendance (claimHex, pubKey, password) {
+  send('log', `attestingClaimOfATtendance with claim: ${claimHex} pubKey: ${pubKey}, pwd: ${password}`);
+  const registry = api.registry;
+  const keyPair = keyring.getPair(hexToU8a(pubKey));
+  try {
+    keyPair.decodePkcs8(password);
+  } catch (err) {
+    return new Promise((resolve, reject) => {
+      resolve({ error: 'password check failed' });
+    });
+  }
+  const publicKey = createType(registry, 'AccountId', pubKey).toU8a();
+  const sign = await api.sign(keyPair, hexToU8a(claimHex));
+
+  const attest = createType(registry, 'Attestation', {
+    claim: claimHex,
+    signature: sign,
+    public: publicKey
+  });
+  return new Promise((resolve, reject) => {
+    resolve({ attestation: attest, attestationHex: u8aToHex(attest.toU8a()) });
+  });
+}
+
 function checkPassword (pubKey, pass) {
   return new Promise((resolve, reject) => {
     const keyPair = keyring.getPair(hexToU8a(pubKey));
@@ -325,5 +360,6 @@ export default {
   sendTx,
   checkPassword,
   changePassword,
-  checkDerivePath
+  checkDerivePath,
+  attestClaimOfAttendance
 };
