+++ lib/js_service_encointer/src/service/encointer.js
@@ -199,41 +199,6 @@ export function parseClaimOfAttendance (claimHex) {
   };
 }
 
-/**
- * Attest Claim Of Attendance with signature
- * @param claimHex
- * @param {AccountId} pubKey
- * @param password
- * @return {U8a} attest
- */
-export async function attestClaimOfAttendance (claimHex, pubKey, password) {
-  const registry = api.registry;
-  await cryptoWaitReady();
-  return new Promise((resolve, reject) => {
-    send('log', `attestingClaimOfATtendance with claim: ${claimHex} pubKey: ${pubKey}, pwd: ${password}`);
-    const keyPair = keyring.getPair(hexToU8a(pubKey));
-    send('log', `got keypair encoded: ${keyPair}`);
-    try {
-      keyPair.decodePkcs8(password);
-    } catch (err) {
-      resolve({ error: 'password check failed' });
-    }
-    send('log', `attestingClaimOfATtendance with claim: ${claimHex} and account: ${pubKey}`);
-
-    const publicKey = createType(registry, 'AccountId', pubKey).toU8a();
-    send('log', `Got public key ${publicKey}`);
-    const sign = pubKey.sign(claimHex);
-    send('log', `Got sign ${sign}`);
-
-    const attest = createType(registry, 'Attestation', {
-      claim: claimHex,
-      signature: sign,
-      public: publicKey
-    });
-    resolve(attest.toU8a());
-  });
-}
-
 export default {
   fetchCurrentPhase,
   fetchCurrencyIdentifiers,
@@ -242,7 +207,6 @@ export default {
   getProofOfAttendance,
   getClaimOfAttendance,
   parseClaimOfAttendance,
-  attestClaimOfAttendance,
   fetchCurrentCeremonyIndex,
   fetchNextMeetupLocation,
   fetchNextMeetupTime,
